When VOD playback starts all torrents in the library are paused. After VOD playback ends all torrents are resumed again.
Before this patch all torrents were resumed, this included the
torrents that the user explicitly set paused herself. This is now
fixed. 

A file maintains the user choices (stop or resume) for each infohash
in the library. Those torrents that are manually stopped will not be
resumed after VOD playback ends.


git-svn-id: http://svn.tribler.org/abc/branches/mainbranch@13982 001aeff7-3401-0410-a489-f7902fc005dd