Problem: - start tribler - download: ubuntu - playback: big bunny - auto pause: ubuntu - close tribler (while playing) - start tribler - big bunny = downloading - ubuntu = paused
This is now fixed.  I had something in place to decrease CPU usage,
namely, only when the user initiated a VOD playback would I look for
torrents to auto-resume.  This was not stored persistently, so after
closing and starting Tribler it would no longer try to auto-resume
(until another VOD playback attempt was started).

Now I will check once after tribler starts.  So an easy fix :) to bad
it takes so long to find :|


git-svn-id: http://svn.tribler.org/abc/branches/mainbranch@14220 001aeff7-3401-0410-a489-f7902fc005dd